+++ build.gradle
@@ -3,28 +3,31 @@ plugins {
     id 'org.jetbrains.kotlin.jvm' version '1.3.21'
 }
 
-group 'com.github.ackintosh'
+//group 'com.github.ackintosh.plasmachain'
 version '1.0-SNAPSHOT'
 
 sourceCompatibility = 1.8
 
-repositories {
-    mavenCentral()
-}
+allprojects {
+    repositories {
+        mavenCentral()
+    }
 
-dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
-    implementation "com.google.guava:guava:27.1-jre"
-    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter', version: '5.4.2'
-}
+    dependencies {
+        implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
+        implementation "com.google.guava:guava:27.1-jre"
+        testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter', version: '5.4.2'
+    }
 
-compileKotlin {
-    kotlinOptions.jvmTarget = "1.8"
-}
-compileTestKotlin {
-    kotlinOptions.jvmTarget = "1.8"
-}
+    compileKotlin {
+        kotlinOptions.jvmTarget = "1.8"
+    }
+    compileTestKotlin {
+        kotlinOptions.jvmTarget = "1.8"
+    }
 
-test {
-    useJUnitPlatform()
+    test {
+        useJUnitPlatform()
+    }
+
 }
