--- src/main/kotlin/com/github/ackintosh/utxo/transaction/CoinbaseData.kt
@@ -1,3 +1,0 @@
-package com.github.ackintosh.utxo.transaction
-
-data class CoinbaseData(private val data: String)
