--- src/main/kotlin/com/github/ackintosh/utxo/transaction/GenerationTransaction.kt
@@ -1,8 +1,0 @@
-package com.github.ackintosh.utxo.transaction
-
-class GenerationTransaction(
-    private val coinbaseData: CoinbaseData
-) {
-    fun transactionHash() = Hash.GENERATION
-    fun outputIndex() = OutputIndex.GENERATION
-}
