--- src/main/kotlin/com/github/ackintosh/utxo/transaction/Input.kt
@@ -1,8 +1,0 @@
-package com.github.ackintosh.utxo.transaction
-
-class Input(
-    val transactionHash: Hash,
-    val outputIndex: OutputIndex
-) {
-    fun toHexString() = "${transactionHash.hash}${outputIndex.toHexString()}"
-}
