--- src/test/kotlin/com/github/ackintosh/utxo/BlockTest.kt
@@ -1,44 +1,0 @@
-package com.github.ackintosh.utxo
-
-import com.github.ackintosh.utxo.merkletree.MerkleTree
-import com.github.ackintosh.utxo.transaction.Transaction
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Test
-
-class BlockTest {
-    @Test
-    fun transactionCounter() {
-        val transactions = listOf(
-            Transaction(inputs = emptyList(), outputs = emptyList()),
-            Transaction(inputs = emptyList(), outputs = emptyList())
-        )
-
-        val block = Block(
-            header = BlockHeader(
-                previousBlockHash = PreviousBlockHash("xxx"),
-                merkleRoot = MerkleTree.build(transactions.map { it.transactionHash() })
-            ),
-            transactions = transactions
-        )
-
-        assertEquals(2, block.transactionCounter())
-    }
-
-    @Test
-    fun blockHash() {
-        val transactions = listOf(
-            Transaction(inputs = emptyList(), outputs = emptyList()),
-            Transaction(inputs = emptyList(), outputs = emptyList())
-        )
-
-        val block = Block(
-            header = BlockHeader(
-                previousBlockHash = PreviousBlockHash("xxx"),
-                merkleRoot = MerkleTree.build(transactions.map { it.transactionHash() })
-            ),
-            transactions = transactions
-        )
-
-        assertEquals("90c5ddbc0a88ea749b48e0c07ff66381abbaddce2f076bb1055112348135bb1a", block.blockHash())
-    }
-}
