--- src/test/kotlin/com/github/ackintosh/utxo/transaction/GenerationTransactionTest.kt
@@ -1,22 +1,0 @@
-package com.github.ackintosh.utxo.transaction
-
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Test
-
-class GenerationTransactionTest {
-    @Test
-    fun transactionHash() {
-        assertEquals(
-            Hash.GENERATION,
-            GenerationTransaction(CoinbaseData("x")).transactionHash()
-        )
-    }
-
-    @Test
-    fun outputIndex() {
-        assertEquals(
-            OutputIndex.GENERATION,
-            GenerationTransaction(CoinbaseData("x")).outputIndex()
-        )
-    }
-}
