+++ utxo/src/main/kotlin/com/github/ackintosh/plasmachain/utxo/transaction/CoinbaseData.kt
@@ -1,0 +1,3 @@
+package com.github.ackintosh.plasmachain.utxo.transaction
+
+data class CoinbaseData(private val data: String)
