+++ utxo/src/main/kotlin/com/github/ackintosh/plasmachain/utxo/transaction/GenerationTransaction.kt
@@ -1,0 +1,8 @@
+package com.github.ackintosh.plasmachain.utxo.transaction
+
+class GenerationTransaction(
+    private val coinbaseData: CoinbaseData
+) {
+    fun transactionHash() = Hash.GENERATION
+    fun outputIndex() = OutputIndex.GENERATION
+}
