+++ utxo/src/main/kotlin/com/github/ackintosh/plasmachain/utxo/transaction/Input.kt
@@ -1,0 +1,8 @@
+package com.github.ackintosh.plasmachain.utxo.transaction
+
+class Input(
+    val transactionHash: Hash,
+    val outputIndex: OutputIndex
+) {
+    fun toHexString() = "${transactionHash.hash}${outputIndex.toHexString()}"
+}
