+++ utxo/src/test/kotlin/com/github/ackintosh/plasmachain/utxo/transaction/OutputIndexTest.kt
@@ -1,0 +1,11 @@
+package com.github.ackintosh.plasmachain.utxo.transaction
+
+import org.junit.jupiter.api.Assertions
+import org.junit.jupiter.api.Test
+
+class OutputIndexTest {
+    @Test
+    fun toHexString() {
+        Assertions.assertEquals("0000000a", OutputIndex(10u).toHexString())
+    }
+}
