+++ utxo/src/test/kotlin/com/github/ackintosh/plasmachain/utxo/transaction/OutputTest.kt
@@ -1,0 +1,11 @@
+package com.github.ackintosh.plasmachain.utxo.transaction
+
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
+
+class OutputTest {
+    @Test
+    fun toHexString() {
+        assertEquals("000000000000000a", Output(10).toHexString())
+    }
+}
